requirements.txt
streamlit==1.25.0
pandas==2.1.1
pyodbc==4.0.40
azure-identity==1.17.0